requirements.txt
asgiref==3.5.2
bleach==5.0.1
captcha==0.4
certifi==2022.6.15.2
charset-normalizer==2.1.1
commonmark==0.9.1
Django==4.1.1
django-filter==22.1
django-restql==0.15.2
djangorestframework==3.12.2
djangorestframework-simplejwt==5.2.0
docutils==0.19
et-xmlfile==1.1.0
idna==3.3
importlib-metadata==4.12.0
jaraco.classes==3.2.2
keyring==23.9.1
more-itertools==8.14.0
openpyxl==3.0.10
Pillow==9.2.0
pkginfo==1.8.3
Pygments==2.13.0
PyJWT==2.4.0
pyPEG2==2.15.2
pypinyin==0.47.1
readme-renderer==37.1
requests==2.28.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==12.5.1
six==1.16.0
sqlparse==0.4.2
twine==4.0.1
ua-parser==0.16.1
urllib3==1.26.12
user-agents==2.2.0
webencodings==0.5.1
zipp==3.8.1
